ARM: OMAP4: add creation of iva sequencer devices for OMAP4
The IVAHD has 2 sequencers in OMAP4. This patch is adding them so that
they can be turn on too.

NOTE: default latency functions does not work for iva sequencers, they
need to be under reset until the IVAHD load some valid code. So, we
are calling hwmod shutdown when they are deactivated.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>